refactor(lambda): extract tests + event_sources/permissions impl groups
service.rs grew to 2935 lines after Phase I extraction. Two-step split:
- Move inline #[cfg(test)] mod tests {} (lines 2098-2935, 835 lines) to
  service_tests.rs sibling.
- Extract event-source-mapping methods (5) and permission/policy methods
  (3) to service_event_sources.rs (268 lines) and service_permissions.rs
  (249 lines) via #[path] modules with impl LambdaService blocks.

service.rs: 2935 -> 1616 lines (-45%).
Methods are pub(super) so service_tests.rs can call them.
