+++ crates/fakecloud-lambda/src/service.rs
@@ -1240,221 +1240,6 @@ impl LambdaService {
         Ok(AwsResponse::json(StatusCode::CREATED, config.to_string()))
     }
 
-    fn create_event_source_mapping(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body: Value = serde_json::from_slice(&req.body).unwrap_or_default();
-        let event_source_arn = body["EventSourceArn"]
-            .as_str()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "InvalidParameterValueException",
-                    "EventSourceArn is required",
-                )
-            })?
-            .to_string();
-
-        let function_name = body["FunctionName"]
-            .as_str()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "InvalidParameterValueException",
-                    "FunctionName is required",
-                )
-            })?
-            .to_string();
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-
-        // Resolve function name to ARN
-        let function_arn = if function_name.starts_with("arn:") {
-            function_name.clone()
-        } else {
-            let func = state.functions.get(&function_name).ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::NOT_FOUND,
-                    "ResourceNotFoundException",
-                    format!(
-                        "Function not found: arn:aws:lambda:{}:{}:function:{}",
-                        state.region, state.account_id, function_name
-                    ),
-                )
-            })?;
-            func.function_arn.clone()
-        };
-
-        let batch_size = body["BatchSize"].as_i64().unwrap_or(10);
-        let enabled = body["Enabled"].as_bool().unwrap_or(true);
-        let mapping_uuid = uuid::Uuid::new_v4().to_string();
-        let now = Utc::now();
-
-        // Extract Filters[].Pattern strictly: any entry where
-        // `Pattern` is missing or not a string is a hard error,
-        // matching AWS. Doing this before `validate` keeps malformed
-        // values from being silently dropped by the lossy serializer.
-        // FilterCriteria itself must be an object (or absent) — non-
-        // object values would otherwise be silently dropped by
-        // `Value::get`, masking client bugs.
-        let filter_patterns: Vec<String> = match body.get("FilterCriteria") {
-            None | Some(Value::Null) => Vec::new(),
-            Some(Value::Object(_)) => {
-                match body.get("FilterCriteria").and_then(|v| v.get("Filters")) {
-                    None => Vec::new(),
-                    Some(Value::Array(arr)) => {
-                        let mut out = Vec::with_capacity(arr.len());
-                        for f in arr {
-                            match f.get("Pattern") {
-                                Some(Value::String(s)) => out.push(s.clone()),
-                                _ => {
-                                    return Err(AwsServiceError::aws_error(
-                                        StatusCode::BAD_REQUEST,
-                                        "InvalidParameterValueException",
-                                        "FilterCriteria.Filters[].Pattern must be a string",
-                                    ));
-                                }
-                            }
-                        }
-                        out
-                    }
-                    Some(_) => {
-                        return Err(AwsServiceError::aws_error(
-                            StatusCode::BAD_REQUEST,
-                            "InvalidParameterValueException",
-                            "FilterCriteria.Filters must be an array",
-                        ));
-                    }
-                }
-            }
-            Some(_) => {
-                return Err(AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "InvalidParameterValueException",
-                    "FilterCriteria must be an object",
-                ));
-            }
-        };
-        // AWS rejects malformed FilterCriteria at create time.
-        if let Err(err) = crate::filter::FilterSet::validate(filter_patterns.iter()) {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "InvalidParameterValueException",
-                err,
-            ));
-        }
-        let function_response_types: Vec<String> = body
-            .get("FunctionResponseTypes")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let starting_position = body
-            .get("StartingPosition")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-        let starting_position_timestamp = body
-            .get("StartingPositionTimestamp")
-            .and_then(|v| v.as_f64());
-        let parallelization_factor = body.get("ParallelizationFactor").and_then(|v| v.as_i64());
-        let maximum_batching_window_in_seconds = body
-            .get("MaximumBatchingWindowInSeconds")
-            .and_then(|v| v.as_i64());
-
-        let mapping = EventSourceMapping {
-            uuid: mapping_uuid.clone(),
-            function_arn: function_arn.clone(),
-            event_source_arn: event_source_arn.clone(),
-            batch_size,
-            enabled,
-            state: if enabled {
-                "Enabled".to_string()
-            } else {
-                "Disabled".to_string()
-            },
-            last_modified: now,
-            filter_patterns,
-            maximum_batching_window_in_seconds,
-            starting_position,
-            starting_position_timestamp,
-            parallelization_factor,
-            function_response_types,
-        };
-
-        let response = self.event_source_mapping_json(&mapping);
-        state.event_source_mappings.insert(mapping_uuid, mapping);
-
-        Ok(AwsResponse::json(
-            StatusCode::ACCEPTED,
-            response.to_string(),
-        ))
-    }
-
-    fn list_event_source_mappings(&self, account_id: &str) -> Result<AwsResponse, AwsServiceError> {
-        let accounts = self.state.read();
-        let empty = LambdaState::new(account_id, "");
-        let state = accounts.get(account_id).unwrap_or(&empty);
-        let mappings: Vec<Value> = state
-            .event_source_mappings
-            .values()
-            .map(|m| self.event_source_mapping_json(m))
-            .collect();
-
-        let response = json!({
-            "EventSourceMappings": mappings,
-        });
-
-        Ok(AwsResponse::json(StatusCode::OK, response.to_string()))
-    }
-
-    fn get_event_source_mapping(
-        &self,
-        uuid: &str,
-        account_id: &str,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let accounts = self.state.read();
-        let empty = LambdaState::new(account_id, "");
-        let state = accounts.get(account_id).unwrap_or(&empty);
-        let mapping = state.event_source_mappings.get(uuid).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::NOT_FOUND,
-                "ResourceNotFoundException",
-                format!("The resource you requested does not exist. (Service: Lambda, Status Code: 404, Request ID: {uuid})"),
-            )
-        })?;
-
-        let response = self.event_source_mapping_json(mapping);
-        Ok(AwsResponse::json(StatusCode::OK, response.to_string()))
-    }
-
-    fn delete_event_source_mapping(
-        &self,
-        uuid: &str,
-        account_id: &str,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(account_id);
-        let mapping = state.event_source_mappings.remove(uuid).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::NOT_FOUND,
-                "ResourceNotFoundException",
-                format!("The resource you requested does not exist. (Service: Lambda, Status Code: 404, Request ID: {uuid})"),
-            )
-        })?;
-
-        let mut response = self.event_source_mapping_json(&mapping);
-        response["State"] = json!("Deleting");
-        Ok(AwsResponse::json(
-            StatusCode::ACCEPTED,
-            response.to_string(),
-        ))
-    }
-
     pub(crate) fn function_config_json(&self, func: &LambdaFunction) -> Value {
         let mut env_vars = json!({});
         if !func.environment.is_empty() {
@@ -1496,282 +1281,6 @@ impl LambdaService {
                 .collect::<Vec<_>>());
         }
         config
-    }
-
-    fn event_source_mapping_json(&self, mapping: &EventSourceMapping) -> Value {
-        let mut out = json!({
-            "UUID": mapping.uuid,
-            "FunctionArn": mapping.function_arn,
-            "EventSourceArn": mapping.event_source_arn,
-            "BatchSize": mapping.batch_size,
-            "State": mapping.state,
-            "LastModified": mapping.last_modified.timestamp_millis() as f64 / 1000.0,
-        });
-        let obj = out.as_object_mut().expect("json! built object");
-        if !mapping.filter_patterns.is_empty() {
-            obj.insert(
-                "FilterCriteria".into(),
-                json!({
-                    "Filters": mapping.filter_patterns.iter().map(|p| json!({"Pattern": p})).collect::<Vec<_>>(),
-                }),
-            );
-        }
-        if !mapping.function_response_types.is_empty() {
-            obj.insert(
-                "FunctionResponseTypes".into(),
-                json!(mapping.function_response_types),
-            );
-        }
-        if let Some(sp) = &mapping.starting_position {
-            obj.insert("StartingPosition".into(), json!(sp));
-        }
-        if let Some(ts) = mapping.starting_position_timestamp {
-            obj.insert("StartingPositionTimestamp".into(), json!(ts));
-        }
-        if let Some(pf) = mapping.parallelization_factor {
-            obj.insert("ParallelizationFactor".into(), json!(pf));
-        }
-        if let Some(w) = mapping.maximum_batching_window_in_seconds {
-            obj.insert("MaximumBatchingWindowInSeconds".into(), json!(w));
-        }
-        out
-    }
-
-    /// Grant a permission on a Lambda function by appending a
-    /// statement to its resource-based policy.
-    ///
-    /// Mirrors AWS: the caller passes `(StatementId, Action,
-    /// Principal, SourceArn?, SourceAccount?)` and the service
-    /// composes a canonical policy document so that the existing
-    /// evaluator can read it without a Lambda-specific fork. Per the
-    /// S3 rollout's #427 evaluator, `SourceArn` becomes an `ArnLike`
-    /// Condition and `SourceAccount` becomes a `StringEquals`
-    /// Condition — both are already supported by the Phase 2 operator
-    /// set, so the permission gate behaves end-to-end without any new
-    /// evaluator code.
-    fn add_permission(
-        &self,
-        function_name: &str,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body: Value = serde_json::from_slice(&req.body).unwrap_or_default();
-        let statement_id = body
-            .get("StatementId")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "InvalidParameterValueException",
-                    "StatementId is required",
-                )
-            })?
-            .to_string();
-        let action = body
-            .get("Action")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "InvalidParameterValueException",
-                    "Action is required",
-                )
-            })?
-            .to_string();
-        let principal_raw = body
-            .get("Principal")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "InvalidParameterValueException",
-                    "Principal is required",
-                )
-            })?
-            .to_string();
-        let source_arn = body
-            .get("SourceArn")
-            .and_then(|v| v.as_str())
-            .map(str::to_string);
-        let source_account = body
-            .get("SourceAccount")
-            .and_then(|v| v.as_str())
-            .map(str::to_string);
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let func = state.functions.get_mut(function_name).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::NOT_FOUND,
-                "ResourceNotFoundException",
-                format!("Function not found: {function_name}"),
-            )
-        })?;
-
-        // Load current policy or seed a fresh canonical doc. Any
-        // stored blob that doesn't parse as a JSON object is treated
-        // as corrupt and replaced — `AddPermission` is the only
-        // mutation path for this field and it always writes valid
-        // JSON, so seeing a non-object here means something else
-        // wrote garbage, and silently propagating it would make
-        // later reads harder to debug.
-        let mut doc: Value = func
-            .policy
-            .as_deref()
-            .and_then(|s| serde_json::from_str::<Value>(s).ok())
-            .filter(|v| v.is_object())
-            .unwrap_or_else(|| json!({"Version": "2012-10-17", "Statement": []}));
-
-        // Ensure Statement is an array so we can push into it.
-        if !doc.get("Statement").map(|s| s.is_array()).unwrap_or(false) {
-            doc["Statement"] = json!([]);
-        }
-        let statements = doc["Statement"].as_array_mut().unwrap();
-
-        // Reject duplicate StatementId — matches AWS's
-        // ResourceConflictException.
-        if statements
-            .iter()
-            .any(|s| s.get("Sid").and_then(|v| v.as_str()) == Some(statement_id.as_str()))
-        {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::CONFLICT,
-                "ResourceConflictException",
-                format!("The statement id ({statement_id}) provided already exists"),
-            ));
-        }
-
-        // Canonicalize Principal: a service host string becomes
-        // `{"Service": "<host>"}`, an account-id or ARN becomes
-        // `{"AWS": "<raw>"}`. AWS accepts both shapes on the wire;
-        // storing the object form uniformly means the existing
-        // evaluator path handles everything without reading back the
-        // raw input.
-        let principal_value =
-            if principal_raw.ends_with(".amazonaws.com") || principal_raw.contains(".amazon") {
-                json!({ "Service": principal_raw })
-            } else {
-                json!({ "AWS": principal_raw })
-            };
-
-        // Emit SourceArn / SourceAccount as Condition keys so the
-        // existing Phase 2 ArnLike / StringEquals operators gate the
-        // grant without new evaluator code.
-        let mut condition = serde_json::Map::new();
-        if let Some(arn) = source_arn.as_ref() {
-            condition.insert("ArnLike".to_string(), json!({ "aws:SourceArn": arn }));
-        }
-        if let Some(acct) = source_account.as_ref() {
-            condition.insert(
-                "StringEquals".to_string(),
-                json!({ "aws:SourceAccount": acct }),
-            );
-        }
-
-        let mut new_statement = serde_json::Map::new();
-        new_statement.insert("Sid".to_string(), json!(statement_id));
-        new_statement.insert("Effect".to_string(), json!("Allow"));
-        new_statement.insert("Principal".to_string(), principal_value);
-        new_statement.insert("Action".to_string(), json!(format!("lambda:{action}")));
-        new_statement.insert("Resource".to_string(), json!(func.function_arn));
-        if !condition.is_empty() {
-            new_statement.insert("Condition".to_string(), Value::Object(condition));
-        }
-        let statement_json = Value::Object(new_statement);
-        statements.push(statement_json.clone());
-
-        func.policy = Some(serde_json::to_string(&doc).unwrap());
-
-        Ok(AwsResponse::json(
-            StatusCode::CREATED,
-            json!({ "Statement": serde_json::to_string(&statement_json).unwrap() }).to_string(),
-        ))
-    }
-
-    fn remove_permission(
-        &self,
-        function_name: &str,
-        statement_id: &str,
-        account_id: &str,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(account_id);
-        let func = state.functions.get_mut(function_name).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::NOT_FOUND,
-                "ResourceNotFoundException",
-                format!("Function not found: {function_name}"),
-            )
-        })?;
-        let policy_str = func.policy.as_deref().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::NOT_FOUND,
-                "ResourceNotFoundException",
-                format!("No policy is associated with function {function_name}"),
-            )
-        })?;
-        let mut doc: Value = serde_json::from_str(policy_str).map_err(|_| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "InternalError",
-                "stored resource policy is not valid JSON",
-            )
-        })?;
-        let statements = doc
-            .get_mut("Statement")
-            .and_then(|s| s.as_array_mut())
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "InternalError",
-                    "stored resource policy has no Statement array",
-                )
-            })?;
-        let before = statements.len();
-        statements.retain(|s| s.get("Sid").and_then(|v| v.as_str()) != Some(statement_id));
-        if statements.len() == before {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::NOT_FOUND,
-                "ResourceNotFoundException",
-                format!("Statement {statement_id} is not found in resource policy"),
-            ));
-        }
-        // Leave an empty {"Statement":[]} behind rather than clearing
-        // the field to None — AWS's GetPolicy keeps returning the
-        // (empty) doc until the function itself is deleted.
-        func.policy = Some(serde_json::to_string(&doc).unwrap());
-        Ok(AwsResponse::json(StatusCode::NO_CONTENT, String::new()))
-    }
-
-    fn get_policy(
-        &self,
-        function_name: &str,
-        account_id: &str,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let accounts = self.state.read();
-        let empty = LambdaState::new(account_id, "");
-        let state = accounts.get(account_id).unwrap_or(&empty);
-        let func = state.functions.get(function_name).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::NOT_FOUND,
-                "ResourceNotFoundException",
-                format!("Function not found: {function_name}"),
-            )
-        })?;
-        let policy = func.policy.as_deref().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::NOT_FOUND,
-                "ResourceNotFoundException",
-                format!("No policy is associated with function {function_name}"),
-            )
-        })?;
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            json!({
-                "Policy": policy,
-                "RevisionId": uuid::Uuid::new_v4().to_string(),
-            })
-            .to_string(),
-        ))
     }
 }
 
@@ -2095,841 +1604,11 @@ impl AwsService for LambdaService {
     }
 }
 
+#[path = "service_event_sources.rs"]
+mod service_event_sources;
+#[path = "service_permissions.rs"]
+mod service_permissions;
+
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use bytes::Bytes;
-    use http::{HeaderMap, Method};
-    use parking_lot::RwLock;
-    use std::collections::HashMap;
-    use std::sync::Arc;
-
-    fn make_state() -> SharedLambdaState {
-        Arc::new(RwLock::new(
-            fakecloud_core::multi_account::MultiAccountState::new("123456789012", "us-east-1", ""),
-        ))
-    }
-
-    fn make_request(method: Method, path: &str, body: &str) -> AwsRequest {
-        let path_segments: Vec<String> = path
-            .split('/')
-            .filter(|s| !s.is_empty())
-            .map(|s| s.to_string())
-            .collect();
-        AwsRequest {
-            service: "lambda".to_string(),
-            action: String::new(),
-            region: "us-east-1".to_string(),
-            account_id: "123456789012".to_string(),
-            request_id: "test-request-id".to_string(),
-            headers: HeaderMap::new(),
-            query_params: HashMap::new(),
-            body: Bytes::from(body.to_string()),
-            body_stream: parking_lot::Mutex::new(None),
-            path_segments,
-            raw_path: path.to_string(),
-            raw_query: String::new(),
-            method,
-            is_query_protocol: false,
-            access_key_id: None,
-            principal: None,
-        }
-    }
-
-    #[test]
-    fn normalize_function_name_bare_name_passes_through() {
-        assert_eq!(normalize_function_name("MyFunction"), "MyFunction");
-    }
-
-    #[test]
-    fn normalize_function_name_strips_qualifier_from_bare_name() {
-        assert_eq!(normalize_function_name("MyFunction:PROD"), "MyFunction");
-        assert_eq!(normalize_function_name("MyFunction:1"), "MyFunction");
-    }
-
-    #[test]
-    fn normalize_function_name_strips_full_arn() {
-        assert_eq!(
-            normalize_function_name("arn:aws:lambda:us-east-1:123456789012:function:MyFunction"),
-            "MyFunction"
-        );
-    }
-
-    #[test]
-    fn normalize_function_name_strips_qualified_full_arn() {
-        assert_eq!(
-            normalize_function_name(
-                "arn:aws:lambda:us-east-1:123456789012:function:MyFunction:PROD"
-            ),
-            "MyFunction"
-        );
-    }
-
-    #[test]
-    fn normalize_function_name_strips_partial_arn() {
-        assert_eq!(
-            normalize_function_name("123456789012:function:MyFunction"),
-            "MyFunction"
-        );
-        assert_eq!(
-            normalize_function_name("123456789012:function:MyFunction:1"),
-            "MyFunction"
-        );
-    }
-
-    #[test]
-    fn normalize_function_name_leaves_malformed_arn_alone() {
-        // wrong service in ARN — multiple colons, no lambda prefix → unchanged
-        let s = "arn:aws:s3:us-east-1:123456789012:function:Foo";
-        assert_eq!(normalize_function_name(s), s);
-        // partial ARN with non-numeric account-shaped prefix → unchanged
-        let s2 = "abc:function:Foo";
-        assert_eq!(normalize_function_name(s2), s2);
-    }
-
-    #[test]
-    fn normalize_function_name_empty() {
-        assert_eq!(normalize_function_name(""), "");
-    }
-
-    #[test]
-    fn normalize_function_name_decodes_percent_encoded_arn() {
-        // SDKs URL-encode `:` in path segments. The toolkit / aws-sdk-lambda
-        // wire form for `arn:aws:lambda:...` is `arn%3Aaws%3Alambda%3A...`.
-        let encoded = "arn%3Aaws%3Alambda%3Aus-east-1%3A123456789012%3Afunction%3AMyFunc";
-        assert_eq!(normalize_function_name(encoded), "MyFunc");
-    }
-
-    #[tokio::test]
-    async fn get_function_accepts_full_arn() {
-        let svc = LambdaService::new(make_state());
-        // Seed a function via CreateFunction
-        let create_body = json!({
-            "FunctionName": "MyFunc",
-            "Runtime": "nodejs20.x",
-            "Role": "arn:aws:iam::123456789012:role/lambda-role",
-            "Handler": "index.handler",
-            "Code": {"ZipFile": ""},
-        })
-        .to_string();
-        let req = make_request(Method::POST, "/2015-03-31/functions", &create_body);
-        svc.handle(req).await.expect("create function");
-
-        // GetFunction by full ARN
-        let req = make_request(
-            Method::GET,
-            "/2015-03-31/functions/arn:aws:lambda:us-east-1:123456789012:function:MyFunc",
-            "",
-        );
-        let resp = svc.handle(req).await.expect("get function by ARN");
-        assert_eq!(resp.status, StatusCode::OK);
-    }
-
-    #[tokio::test]
-    async fn get_function_accepts_partial_arn() {
-        let svc = LambdaService::new(make_state());
-        let create_body = json!({
-            "FunctionName": "MyFunc",
-            "Runtime": "nodejs20.x",
-            "Role": "arn:aws:iam::123456789012:role/lambda-role",
-            "Handler": "index.handler",
-            "Code": {"ZipFile": ""},
-        })
-        .to_string();
-        let req = make_request(Method::POST, "/2015-03-31/functions", &create_body);
-        svc.handle(req).await.expect("create function");
-
-        let req = make_request(
-            Method::GET,
-            "/2015-03-31/functions/123456789012:function:MyFunc",
-            "",
-        );
-        let resp = svc.handle(req).await.expect("get function by partial ARN");
-        assert_eq!(resp.status, StatusCode::OK);
-    }
-
-    #[tokio::test]
-    async fn get_function_accepts_name_with_qualifier() {
-        let svc = LambdaService::new(make_state());
-        let create_body = json!({
-            "FunctionName": "MyFunc",
-            "Runtime": "nodejs20.x",
-            "Role": "arn:aws:iam::123456789012:role/lambda-role",
-            "Handler": "index.handler",
-            "Code": {"ZipFile": ""},
-        })
-        .to_string();
-        let req = make_request(Method::POST, "/2015-03-31/functions", &create_body);
-        svc.handle(req).await.expect("create function");
-
-        let req = make_request(Method::GET, "/2015-03-31/functions/MyFunc:1", "");
-        let resp = svc
-            .handle(req)
-            .await
-            .expect("get function by name:qualifier");
-        assert_eq!(resp.status, StatusCode::OK);
-    }
-
-    #[test]
-    fn iam_condition_keys_for_add_permission_populates_arn_and_principal() {
-        let svc = LambdaService::new(make_state());
-        let body = json!({
-            "StatementId": "stmt",
-            "Action": "lambda:InvokeFunction",
-            "Principal": "s3.amazonaws.com",
-        })
-        .to_string();
-        let req = make_request(Method::POST, "/2015-03-31/functions/my-func/policy", &body);
-        let action = fakecloud_core::auth::IamAction {
-            service: "lambda",
-            action: "AddPermission",
-            resource: "arn:aws:lambda:us-east-1:123456789012:function:my-func".to_string(),
-        };
-        let keys = svc.iam_condition_keys_for(&req, &action);
-        assert_eq!(
-            keys.get("lambda:functionarn"),
-            Some(&vec![
-                "arn:aws:lambda:us-east-1:123456789012:function:my-func".to_string()
-            ])
-        );
-        assert_eq!(
-            keys.get("lambda:principal"),
-            Some(&vec!["s3.amazonaws.com".to_string()])
-        );
-    }
-
-    #[test]
-    fn iam_condition_keys_for_add_permission_omits_missing_principal() {
-        let svc = LambdaService::new(make_state());
-        let body = json!({"StatementId": "stmt", "Action": "lambda:InvokeFunction"}).to_string();
-        let req = make_request(Method::POST, "/2015-03-31/functions/my-func/policy", &body);
-        let action = fakecloud_core::auth::IamAction {
-            service: "lambda",
-            action: "AddPermission",
-            resource: "arn:aws:lambda:us-east-1:123456789012:function:my-func".to_string(),
-        };
-        let keys = svc.iam_condition_keys_for(&req, &action);
-        assert!(!keys.contains_key("lambda:principal"));
-        assert!(keys.contains_key("lambda:functionarn"));
-    }
-
-    #[test]
-    fn iam_condition_keys_for_non_add_permission_is_empty() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(Method::GET, "/2015-03-31/functions/my-func", "");
-        let action = fakecloud_core::auth::IamAction {
-            service: "lambda",
-            action: "GetFunction",
-            resource: "arn:aws:lambda:us-east-1:123456789012:function:my-func".to_string(),
-        };
-        assert!(svc.iam_condition_keys_for(&req, &action).is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_create_and_get_function() {
-        let state = make_state();
-        let svc = LambdaService::new(state);
-
-        let create_body = json!({
-            "FunctionName": "my-func",
-            "Runtime": "python3.12",
-            "Role": "arn:aws:iam::123456789012:role/test-role",
-            "Handler": "index.handler",
-            "Code": { "ZipFile": "UEsFBgAAAAAAAAAAAAAAAAAAAAA=" }
-        });
-
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions",
-            &create_body.to_string(),
-        );
-        let resp = svc.handle(req).await.unwrap();
-        assert_eq!(resp.status, StatusCode::CREATED);
-
-        let body: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(body["FunctionName"], "my-func");
-        assert_eq!(body["Runtime"], "python3.12");
-
-        // Get
-        let req = make_request(Method::GET, "/2015-03-31/functions/my-func", "");
-        let resp = svc.handle(req).await.unwrap();
-        assert_eq!(resp.status, StatusCode::OK);
-        let body: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(body["Configuration"]["FunctionName"], "my-func");
-    }
-
-    #[tokio::test]
-    async fn test_delete_function() {
-        let state = make_state();
-        let svc = LambdaService::new(state);
-
-        let create_body = json!({
-            "FunctionName": "to-delete",
-            "Runtime": "nodejs20.x",
-            "Role": "arn:aws:iam::123456789012:role/test",
-            "Handler": "index.handler",
-            "Code": {}
-        });
-
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions",
-            &create_body.to_string(),
-        );
-        svc.handle(req).await.unwrap();
-
-        let req = make_request(Method::DELETE, "/2015-03-31/functions/to-delete", "");
-        let resp = svc.handle(req).await.unwrap();
-        assert_eq!(resp.status, StatusCode::NO_CONTENT);
-
-        // Verify deleted
-        let req = make_request(Method::GET, "/2015-03-31/functions/to-delete", "");
-        let resp = svc.handle(req).await;
-        assert!(resp.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_invoke_without_runtime_returns_error() {
-        let state = make_state();
-        let svc = LambdaService::new(state);
-
-        let create_body = json!({
-            "FunctionName": "invoke-me",
-            "Runtime": "python3.12",
-            "Role": "arn:aws:iam::123456789012:role/test",
-            "Handler": "index.handler",
-            "Code": {}
-        });
-
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions",
-            &create_body.to_string(),
-        );
-        svc.handle(req).await.unwrap();
-
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions/invoke-me/invocations",
-            r#"{"key": "value"}"#,
-        );
-        let resp = svc.handle(req).await;
-        assert!(resp.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_invoke_nonexistent_function() {
-        let state = make_state();
-        let svc = LambdaService::new(state);
-
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions/does-not-exist/invocations",
-            "{}",
-        );
-        let resp = svc.handle(req).await;
-        assert!(resp.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_list_functions() {
-        let state = make_state();
-        let svc = LambdaService::new(state);
-
-        for name in &["func-a", "func-b"] {
-            let create_body = json!({
-                "FunctionName": name,
-                "Runtime": "python3.12",
-                "Role": "arn:aws:iam::123456789012:role/test",
-                "Handler": "index.handler",
-                "Code": {}
-            });
-            let req = make_request(
-                Method::POST,
-                "/2015-03-31/functions",
-                &create_body.to_string(),
-            );
-            svc.handle(req).await.unwrap();
-        }
-
-        let req = make_request(Method::GET, "/2015-03-31/functions", "");
-        let resp = svc.handle(req).await.unwrap();
-        let body: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(body["Functions"].as_array().unwrap().len(), 2);
-    }
-
-    #[tokio::test]
-    async fn test_event_source_mapping() {
-        let state = make_state();
-        let svc = LambdaService::new(state);
-
-        // Create function first
-        let create_body = json!({
-            "FunctionName": "esm-func",
-            "Runtime": "python3.12",
-            "Role": "arn:aws:iam::123456789012:role/test",
-            "Handler": "index.handler",
-            "Code": {}
-        });
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions",
-            &create_body.to_string(),
-        );
-        svc.handle(req).await.unwrap();
-
-        // Create mapping
-        let mapping_body = json!({
-            "FunctionName": "esm-func",
-            "EventSourceArn": "arn:aws:sqs:us-east-1:123456789012:my-queue",
-            "BatchSize": 5
-        });
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/event-source-mappings",
-            &mapping_body.to_string(),
-        );
-        let resp = svc.handle(req).await.unwrap();
-        assert_eq!(resp.status, StatusCode::ACCEPTED);
-        let body: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let uuid = body["UUID"].as_str().unwrap().to_string();
-
-        // List mappings
-        let req = make_request(Method::GET, "/2015-03-31/event-source-mappings", "");
-        let resp = svc.handle(req).await.unwrap();
-        let body: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(body["EventSourceMappings"].as_array().unwrap().len(), 1);
-
-        // Delete mapping
-        let req = make_request(
-            Method::DELETE,
-            &format!("/2015-03-31/event-source-mappings/{uuid}"),
-            "",
-        );
-        let resp = svc.handle(req).await.unwrap();
-        assert_eq!(resp.status, StatusCode::ACCEPTED);
-    }
-
-    async fn seed_function(svc: &LambdaService, name: &str) {
-        let body = json!({
-            "FunctionName": name,
-            "Runtime": "python3.12",
-            "Role": "arn:aws:iam::123456789012:role/r",
-            "Handler": "index.handler",
-            "Code": {}
-        });
-        let req = make_request(Method::POST, "/2015-03-31/functions", &body.to_string());
-        svc.handle(req).await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn add_permission_builds_canonical_statement() {
-        let svc = LambdaService::new(make_state());
-        seed_function(&svc, "f").await;
-
-        let body = json!({
-            "StatementId": "s3-invoke",
-            "Action": "InvokeFunction",
-            "Principal": "s3.amazonaws.com",
-            "SourceArn": "arn:aws:s3:::my-bucket",
-            "SourceAccount": "123456789012",
-        });
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions/f/policy",
-            &body.to_string(),
-        );
-        let resp = svc.handle(req).await.unwrap();
-        assert_eq!(resp.status, StatusCode::CREATED);
-
-        let out: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let statement: Value = serde_json::from_str(out["Statement"].as_str().unwrap()).unwrap();
-        assert_eq!(statement["Sid"], "s3-invoke");
-        assert_eq!(statement["Effect"], "Allow");
-        assert_eq!(statement["Principal"]["Service"], "s3.amazonaws.com");
-        assert_eq!(statement["Action"], "lambda:InvokeFunction");
-        assert_eq!(
-            statement["Resource"],
-            "arn:aws:lambda:us-east-1:123456789012:function:f"
-        );
-        assert_eq!(
-            statement["Condition"]["ArnLike"]["aws:SourceArn"],
-            "arn:aws:s3:::my-bucket"
-        );
-        assert_eq!(
-            statement["Condition"]["StringEquals"]["aws:SourceAccount"],
-            "123456789012"
-        );
-    }
-
-    #[tokio::test]
-    async fn add_permission_aws_principal_emits_aws_key() {
-        let svc = LambdaService::new(make_state());
-        seed_function(&svc, "f").await;
-
-        let body = json!({
-            "StatementId": "user-invoke",
-            "Action": "InvokeFunction",
-            "Principal": "arn:aws:iam::123456789012:user/alice",
-        });
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions/f/policy",
-            &body.to_string(),
-        );
-        svc.handle(req).await.unwrap();
-
-        // Fetch via GetPolicy and inspect the stored doc.
-        let req = make_request(Method::GET, "/2015-03-31/functions/f/policy", "");
-        let resp = svc.handle(req).await.unwrap();
-        let body: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let doc: Value = serde_json::from_str(body["Policy"].as_str().unwrap()).unwrap();
-        let statements = doc["Statement"].as_array().unwrap();
-        assert_eq!(statements.len(), 1);
-        assert_eq!(
-            statements[0]["Principal"]["AWS"],
-            "arn:aws:iam::123456789012:user/alice"
-        );
-        assert!(statements[0].get("Condition").is_none());
-    }
-
-    #[tokio::test]
-    async fn add_permission_rejects_duplicate_statement_id() {
-        let svc = LambdaService::new(make_state());
-        seed_function(&svc, "f").await;
-
-        let body = json!({
-            "StatementId": "dup",
-            "Action": "InvokeFunction",
-            "Principal": "arn:aws:iam::123456789012:user/a",
-        });
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions/f/policy",
-            &body.to_string(),
-        );
-        svc.handle(req).await.unwrap();
-
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions/f/policy",
-            &body.to_string(),
-        );
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-    }
-
-    #[tokio::test]
-    async fn get_policy_returns_404_when_no_policy_attached() {
-        let svc = LambdaService::new(make_state());
-        seed_function(&svc, "f").await;
-
-        let req = make_request(Method::GET, "/2015-03-31/functions/f/policy", "");
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn remove_permission_strips_matching_sid_and_leaves_empty_doc() {
-        let svc = LambdaService::new(make_state());
-        seed_function(&svc, "f").await;
-
-        for sid in ["a", "b"] {
-            let body = json!({
-                "StatementId": sid,
-                "Action": "InvokeFunction",
-                "Principal": "arn:aws:iam::123456789012:user/u",
-            });
-            let req = make_request(
-                Method::POST,
-                "/2015-03-31/functions/f/policy",
-                &body.to_string(),
-            );
-            svc.handle(req).await.unwrap();
-        }
-
-        // Remove "a"
-        let req = make_request(Method::DELETE, "/2015-03-31/functions/f/policy/a", "");
-        let resp = svc.handle(req).await.unwrap();
-        assert_eq!(resp.status, StatusCode::NO_CONTENT);
-
-        // GetPolicy still returns the doc with just "b".
-        let req = make_request(Method::GET, "/2015-03-31/functions/f/policy", "");
-        let resp = svc.handle(req).await.unwrap();
-        let body: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let doc: Value = serde_json::from_str(body["Policy"].as_str().unwrap()).unwrap();
-        let stmts = doc["Statement"].as_array().unwrap();
-        assert_eq!(stmts.len(), 1);
-        assert_eq!(stmts[0]["Sid"], "b");
-
-        // Remove the last one — doc stays (empty Statement array).
-        let req = make_request(Method::DELETE, "/2015-03-31/functions/f/policy/b", "");
-        svc.handle(req).await.unwrap();
-
-        let req = make_request(Method::GET, "/2015-03-31/functions/f/policy", "");
-        let resp = svc.handle(req).await.unwrap();
-        let body: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let doc: Value = serde_json::from_str(body["Policy"].as_str().unwrap()).unwrap();
-        assert_eq!(doc["Statement"].as_array().unwrap().len(), 0);
-    }
-
-    #[tokio::test]
-    async fn remove_permission_unknown_sid_is_404() {
-        let svc = LambdaService::new(make_state());
-        seed_function(&svc, "f").await;
-
-        let body = json!({
-            "StatementId": "known",
-            "Action": "InvokeFunction",
-            "Principal": "arn:aws:iam::123456789012:user/u",
-        });
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions/f/policy",
-            &body.to_string(),
-        );
-        svc.handle(req).await.unwrap();
-
-        let req = make_request(Method::DELETE, "/2015-03-31/functions/f/policy/other", "");
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn add_permission_on_missing_function_is_404() {
-        let svc = LambdaService::new(make_state());
-        let body = json!({
-            "StatementId": "s",
-            "Action": "InvokeFunction",
-            "Principal": "arn:aws:iam::123456789012:user/u",
-        });
-        let req = make_request(
-            Method::POST,
-            "/2015-03-31/functions/missing/policy",
-            &body.to_string(),
-        );
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[test]
-    fn iam_action_for_maps_invoke_to_function_arn() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(Method::POST, "/2015-03-31/functions/f/invocations", "");
-        let action = svc.iam_action_for(&req).unwrap();
-        assert_eq!(action.service, "lambda");
-        assert_eq!(action.action, "InvokeFunction");
-        assert_eq!(
-            action.resource,
-            "arn:aws:lambda:us-east-1:123456789012:function:f"
-        );
-    }
-
-    #[test]
-    fn iam_action_for_maps_list_to_star() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(Method::GET, "/2015-03-31/functions", "");
-        let action = svc.iam_action_for(&req).unwrap();
-        assert_eq!(action.action, "ListFunctions");
-        assert_eq!(action.resource, "*");
-    }
-
-    #[test]
-    fn iam_action_for_create_reads_function_name_from_body() {
-        let svc = LambdaService::new(make_state());
-        let body = json!({
-            "FunctionName": "newfn",
-            "Runtime": "python3.12",
-            "Role": "arn:aws:iam::123456789012:role/r",
-            "Handler": "index.handler",
-            "Code": {}
-        });
-        let req = make_request(Method::POST, "/2015-03-31/functions", &body.to_string());
-        let action = svc.iam_action_for(&req).unwrap();
-        assert_eq!(action.action, "CreateFunction");
-        assert_eq!(
-            action.resource,
-            "arn:aws:lambda:us-east-1:123456789012:function:newfn"
-        );
-    }
-
-    // ── Error branch tests ──
-
-    #[tokio::test]
-    async fn create_function_duplicate_returns_conflict() {
-        let svc = LambdaService::new(make_state());
-        seed_function(&svc, "dup-fn").await;
-
-        let body = json!({
-            "FunctionName": "dup-fn",
-            "Runtime": "python3.12",
-            "Role": "arn:aws:iam::123456789012:role/r",
-            "Handler": "index.handler",
-            "Code": {"ZipFile": "UEsDBBQ="},
-        });
-        let req = make_request(Method::POST, "/2015-03-31/functions", &body.to_string());
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected ResourceConflictException"),
-        };
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-    }
-
-    #[tokio::test]
-    async fn get_function_not_found() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(Method::GET, "/2015-03-31/functions/nope", "");
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn delete_function_not_found() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(Method::DELETE, "/2015-03-31/functions/nope", "");
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn get_event_source_mapping_not_found() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(
-            Method::GET,
-            "/2015-03-31/event-source-mappings/nonexistent",
-            "",
-        );
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn delete_event_source_mapping_not_found() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(
-            Method::DELETE,
-            "/2015-03-31/event-source-mappings/nonexistent",
-            "",
-        );
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn get_policy_on_missing_function() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(Method::GET, "/2015-03-31/functions/nope/policy", "");
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn remove_permission_on_missing_function() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(
-            Method::DELETE,
-            "/2015-03-31/functions/nope/policy/stmt1",
-            "",
-        );
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn publish_version_on_missing_function() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(Method::POST, "/2015-03-31/functions/nope/versions", "{}");
-        let err = match svc.handle(req).await {
-            Err(e) => e,
-            Ok(_) => panic!("expected error"),
-        };
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn unknown_route_returns_error() {
-        let svc = LambdaService::new(make_state());
-        let req = make_request(Method::POST, "/unknown/route", "{}");
-        assert!(svc.handle(req).await.is_err());
-    }
-
-    #[tokio::test]
-    async fn publish_version_unknown_function_errors() {
-        let svc = LambdaService::new(make_state());
-        assert!(svc.publish_version("ghost", "123456789012").is_err());
-    }
-
-    #[tokio::test]
-    async fn get_function_unknown_errors() {
-        let svc = LambdaService::new(make_state());
-        assert!(svc
-            .get_function("ghost", "123456789012", "us-east-1")
-            .is_err());
-    }
-
-    #[tokio::test]
-    async fn delete_function_unknown_errors() {
-        let svc = LambdaService::new(make_state());
-        assert!(svc.delete_function("ghost", "123456789012").is_err());
-    }
-
-    #[tokio::test]
-    async fn get_event_source_mapping_unknown_errors() {
-        let svc = LambdaService::new(make_state());
-        assert!(svc
-            .get_event_source_mapping("ghost", "123456789012")
-            .is_err());
-    }
-
-    #[tokio::test]
-    async fn delete_event_source_mapping_unknown_errors() {
-        let svc = LambdaService::new(make_state());
-        assert!(svc
-            .delete_event_source_mapping("ghost", "123456789012")
-            .is_err());
-    }
-
-    #[tokio::test]
-    async fn list_functions_empty_ok() {
-        let svc = LambdaService::new(make_state());
-        let resp = svc.list_functions("123456789012").unwrap();
-        assert_eq!(resp.status, http::StatusCode::OK);
-    }
-
-    #[tokio::test]
-    async fn list_event_source_mappings_empty_ok() {
-        let svc = LambdaService::new(make_state());
-        let resp = svc.list_event_source_mappings("123456789012").unwrap();
-        assert_eq!(resp.status, http::StatusCode::OK);
-    }
-}
+#[path = "service_tests.rs"]
+mod tests;
